Feat : "Add function to get coordinate when state, city, town were given as a parameter"
Feat : "Add function to handle API data"

+++ src/main/java/com/weatherproject/weather/domain/DTO/ApiUrlGeneratorDTO.java
@@ -19,6 +19,6 @@ public class ApiUrlGeneratorDTO {
     private int baseDate;
     private int baseTime;
     private String dataType;
-    private int nx;
-    private int ny;
+    private long nx;
+    private long ny;
 }
